+++ src/routes/server/examples/hello-world/Main.svelte
@@ -1,7 +1,12 @@
 <script lang="ts">
     import { TableHandler, Datatable, ThSort, RowCount, RowsPerPage, Pagination, type State } from '$lib/src/server'
     import { reload } from './api'
-    let { data }: { data: any } = $props()
+    let { data }: { data: {
+        id: string,
+        name: string,
+        email: string,
+        body: string
+    }[] } = $props()
 
     const table = new TableHandler(data, { rowsPerPage: 5, totalRows: 500 })